backend/requirements.txt
aws-xray-sdk==2.7.0
bidict==0.21.2
boto3==1.17.53
botocore==1.20.53
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
ecdsa==0.14.1
enum34==1.1.10
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT==0.3.2
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jose==1.0.0
MarkupSafe==1.1.1
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.4.2
python-dateutil==2.8.1
python-dotenv==0.17.0
python-engineio==4.1.0
python-jose==3.2.0
python-slugify==4.0.1
python-socketio==5.2.1
requests==2.25.1
rsa==4.7.2
s3transfer==0.3.7
six==1.15.0
text-unidecode==1.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.4.0
